Update mod_pagespeed to be able to build against newer Chromium libraries (r161115), but don't flip over the switch yet, as we also need libpagespeed to be updated, which it should be able to do after this is in.
The bulk of the change is in needing to indirect to scoped_ptr.h as it's
now in a new place; with an addition of a version check for stl_util.h
rename and some build system adjustments, most notable of which is passing
of Chromium version from gclient to gyp to a #define.

Most of the work for this change was done by Libo Song (lsong@)

+++ src/net/instaweb/util/string_util.cc
@@ -17,6 +17,7 @@
 // Author: jmarantz@google.com (Joshua Marantz)
 
 #include "net/instaweb/util/public/string_util.h"
+
 #include <algorithm>
 #include <cctype>
 #include <cstddef>
@@ -327,7 +328,7 @@ int FindIgnoreCase(StringPiece haystack, StringPiece needle) {
     ++pos;
     haystack.remove_prefix(1);
   }
-  return StringPiece::npos;
+  return static_cast<int>(StringPiece::npos);
 }
